Exam_Quiz_Maker_Agent_Using_Chainlit/pyproject.toml
[project]
name = "exam-quiz-maker-agent-using-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
]
